Reorganizing reusable components
