+++ src/components/common/Breadcrumb.tsx
@@ -3,7 +3,7 @@
 import { Box, Breadcrumbs, Typography } from '@mui/material';
 import { usePathname } from 'next/navigation';
 import { useMemo } from 'react';
-import NavLink from './NavLink';
+import NavLink from '../navigation/NavLink';
 
 import { ChevronRightIcon, FileIcon, HomeIcon, SettingsIcon, UserIcon, UsersIcon } from '@/icons';
