+++ src/components/loaders/CustomCircularProgress.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Box, CircularProgress, CircularProgressProps, Typography } from '@mui/material';
 import { useEffect } from 'react';
